task-01-Chatbot with memory/requirements.txt
streamlit
langchain
google-generativeai
langchain-google-genai
python-dotenv